fresccooo

+++ ft_strlen.c
@@ -12,12 +12,12 @@
 
 #include "libft.h"
 
-int ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	unsigned int i;
+	size_t	i;
 
 	i = 0;
-	while(str[i])
+	while (str[i])
 		i++;
 	return (i);
 }
